Moved all test fns to header.
