+++ R/cpp20.R
@@ -1,93 +1,5 @@
 # Generated by cpp20: do not edit by hand
 
-cpp_test <- function(x) {
-  .Call(`_cpp20_cpp_test`, x)
-}
-
-cpp_set_threads <- function(n) {
-  invisible(.Call(`_cpp20_cpp_set_threads`, n))
-}
-
-cpp_get_threads <- function() {
-  .Call(`_cpp20_cpp_get_threads`)
-}
-
-test_null <- function() {
-  .Call(`_cpp20_test_null`)
-}
-
-test1 <- function(n) {
-  .Call(`_cpp20_test1`, n)
-}
-
-test2 <- function(x) {
-  .Call(`_cpp20_test2`, x)
-}
-
-test3 <- function(x) {
-  .Call(`_cpp20_test3`, x)
-}
-
-test4 <- function(x) {
-  .Call(`_cpp20_test4`, x)
-}
-
-test_mix <- function(a, b, c, d) {
-  .Call(`_cpp20_test_mix`, a, b, c, d)
-}
-
-test_mix1 <- function(a, b, c, d, e, f) {
-  .Call(`_cpp20_test_mix1`, a, b, c, d, e, f)
-}
-
-test_sym <- function(x) {
-  .Call(`_cpp20_test_sym`, x)
-}
-
-test_sexp2 <- function(x) {
-  .Call(`_cpp20_test_sexp2`, x)
-}
-
-test_sexp3 <- function(x) {
-  .Call(`_cpp20_test_sexp3`, x)
-}
-
-test_list_to_scalars <- function(x) {
-  .Call(`_cpp20_test_list_to_scalars`, x)
-}
-
-test_coerce1 <- function(x) {
-  .Call(`_cpp20_test_coerce1`, x)
-}
-
-test_construction <- function(x) {
-  .Call(`_cpp20_test_construction`, x)
-}
-
-test_constructions <- function(x) {
-  .Call(`_cpp20_test_constructions`, x)
-}
-
-test_constructions2 <- function(x) {
-  .Call(`_cpp20_test_constructions2`, x)
-}
-
-test_constructions3 <- function(x) {
-  .Call(`_cpp20_test_constructions3`, x)
-}
-
-test_constructions4 <- function(x) {
-  .Call(`_cpp20_test_constructions4`, x)
-}
-
-test_set_strs <- function(x) {
-  .Call(`_cpp20_test_set_strs`, x)
-}
-
-test_set_strs2 <- function(x) {
-  .Call(`_cpp20_test_set_strs2`, x)
-}
-
 test_multiple_deduction <- function(x, y) {
   .Call(`_cpp20_test_multiple_deduction`, x, y)
 }
@@ -106,6 +18,10 @@ test_deduced_scalar_type <- function(x) {
 
 test_identity <- function(x) {
   .Call(`_cpp20_test_identity`, x)
+}
+
+test_identity2 <- function(x) {
+  .Call(`_cpp20_test_identity2`, x)
 }
 
 test_template_null <- function(x) {
@@ -206,4 +122,40 @@ test_specialisation <- function(x) {
 
 test_coerce <- function(x, ptype) {
   .Call(`_cpp20_test_coerce`, x, ptype)
+}
+
+cpp_set_threads <- function(n) {
+  invisible(.Call(`_cpp20_cpp_set_threads`, n))
+}
+
+cpp_get_threads <- function() {
+  .Call(`_cpp20_cpp_get_threads`)
+}
+
+test_null <- function() {
+  .Call(`_cpp20_test_null`)
+}
+
+test4 <- function(x) {
+  .Call(`_cpp20_test4`, x)
+}
+
+test_sym <- function(x) {
+  .Call(`_cpp20_test_sym`, x)
+}
+
+test_sexp2 <- function(x) {
+  .Call(`_cpp20_test_sexp2`, x)
+}
+
+test_sexp3 <- function(x) {
+  .Call(`_cpp20_test_sexp3`, x)
+}
+
+test_list_to_scalars <- function(x) {
+  .Call(`_cpp20_test_list_to_scalars`, x)
+}
+
+test_coerce1 <- function(x) {
+  .Call(`_cpp20_test_coerce1`, x)
 }
